+++ include/JSystem/JKernel/JKRDvdAramRipper.h
@@ -25,7 +25,7 @@ public:
     /* 0x34 */ int field_0x34;
     /* 0x38 */ int field_0x38;
     /* 0x3C */ int field_0x3c;
-    /* 0x40 */ void (*mCallback)(u32);
+    /* 0x40 */ void (*mCallback)(uintptr_t);
     /* 0x44 */ int field_0x44;
     /* 0x48 */ bool field_0x48;
     /* 0x4C */ JKRAramStreamCommand* mStreamCommand;
@@ -36,7 +36,7 @@ class JKRDvdAramRipper {
 public:
     static JKRAramBlock* loadToAram(s32, u32, JKRExpandSwitch, u32, u32);
     static JKRAramBlock* loadToAram(JKRDvdFile*, u32, JKRExpandSwitch, u32, u32);
-    static JKRADCommand* loadToAram_Async(JKRDvdFile*, u32, JKRExpandSwitch, void (*)(u32), u32,
+    static JKRADCommand* loadToAram_Async(JKRDvdFile*, u32, JKRExpandSwitch, void (*)(uintptr_t), u32,
                                           u32);
     static JKRADCommand* callCommand_Async(JKRADCommand*);
     static bool syncAram(JKRADCommand*, int);
