+++ src/JSystem/JKernel/JKRAram.cpp
@@ -102,12 +102,12 @@ void* JKRAram::run() {
 
 /* 802B45D4-802B4664       .text checkOkAddress__7JKRAramFPUcUlP12JKRAramBlockUl */
 bool JKRAram::checkOkAddress(u8* addr, u32 size, JKRAramBlock* block, u32 param_4) {
-    if (!IS_ALIGNED((u32)addr, 0x20) && !IS_ALIGNED(size, 0x20)) {
+    if (!IS_ALIGNED((uintptr_t)addr, 0x20) && !IS_ALIGNED(size, 0x20)) {
         OSPanic(__FILE__, 226, ":::address not 32Byte aligned.");
         return false;
     }
 
-    if (block && !IS_ALIGNED((u32)block->getAddress() + param_4, 0x20)) {
+    if (block && !IS_ALIGNED((uintptr_t)block->getAddress() + param_4, 0x20)) {
         OSPanic(__FILE__, 235, ":::address not 32Byte aligned.");
         return false;
     }
@@ -169,7 +169,7 @@ JKRAramBlock* JKRAram::mainRamToAram(u8 *buf, u32 bufSize, u32 alignedSize, JKRE
         }
         else {
             JKRDecompress(buf, (u8 *)allocatedMem, fileSize, 0);
-            JKRAramPcs(0, (u32)allocatedMem, bufSize, alignedSize, block);
+            JKRAramPcs(0, (uintptr_t)allocatedMem, bufSize, alignedSize, block);
             JKRFreeToHeap(heap, allocatedMem);
             block = block == NULL ? (JKRAramBlock *)-1 : block;
         }
@@ -186,7 +186,7 @@ JKRAramBlock* JKRAram::mainRamToAram(u8 *buf, u32 bufSize, u32 alignedSize, JKRE
             bufSize = allocatedBlock->getAddress();
         }
 
-        JKRAramPcs(0, (u32)buf, bufSize, alignedSize, block);
+        JKRAramPcs(0, (uintptr_t)buf, bufSize, alignedSize, block);
         block = block == NULL ? (JKRAramBlock *)-1 : block;
     }
     return block;
@@ -216,8 +216,8 @@ u8* JKRAram::aramToMainRam(u32 address, u8 *buf, u32 p3, JKRExpandSwitch expandS
     u32 expandSize;
     if (expandSwitch == EXPAND_SWITCH_UNKNOWN1) {
         u8 buffer[64];
-        u8 *bufPtr = (u8 *)ALIGN_NEXT((u32)buffer, 32);
-        JKRAramPcs(1, address, (u32)bufPtr, sizeof(buffer) / 2, NULL); // probably change sizeof(buffer) / 2 to 32
+        u8 *bufPtr = (u8 *)ALIGN_NEXT((uintptr_t)buffer, 32);
+        JKRAramPcs(1, address, (uintptr_t)bufPtr, sizeof(buffer) / 2, NULL); // probably change sizeof(buffer) / 2 to 32
         compression = JKRCheckCompressed(bufPtr);
         expandSize = JKRDecompExpandSize(bufPtr);
     }
@@ -245,7 +245,7 @@ u8* JKRAram::aramToMainRam(u32 address, u8 *buf, u32 p3, JKRExpandSwitch expandS
             return NULL;
         }
         else {
-            JKRAramPcs(1, address, (u32)szpSpace, p3, NULL);
+            JKRAramPcs(1, address, (uintptr_t)szpSpace, p3, NULL);
             if (p5 != 0 && p5 < expandSize)
                 expandSize = p5;
 
@@ -279,7 +279,7 @@ u8* JKRAram::aramToMainRam(u32 address, u8 *buf, u32 p3, JKRExpandSwitch expandS
         }
         else {
             changeGroupIdIfNeed(buf, id);
-            JKRAramPcs(1, address, (u32)buf, p3, NULL);
+            JKRAramPcs(1, address, (uintptr_t)buf, p3, NULL);
             if (pSize != NULL) {
                 *pSize = p3;
             }
@@ -512,8 +512,8 @@ static u8* firstSrcData() {
         length = size;
     }
 
-    u32 src = (u32)(srcAddress + srcOffset);
-    u32 dst = (u32)buffer;
+    uintptr_t src = (uintptr_t)(srcAddress + srcOffset);
+    uintptr_t dst = (uintptr_t)buffer;
     u32 alignedLength = ALIGN_NEXT(length, 0x20);
     JKRAramPcs(1, src, dst, alignedLength, NULL);
 
@@ -545,7 +545,7 @@ static u8* nextSrcData(u8* current) {
     }
     JUT_ASSERT(VERSION_SELECT(1321, 1376, 1361, 1361), transSize > 0);
 
-    JKRAramPcs(1, (u32)(srcAddress + srcOffset), ((u32)dest + left), ALIGN_NEXT(transSize, 0x20),
+    JKRAramPcs(1, (uintptr_t)(srcAddress + srcOffset), ((uintptr_t)dest + left), ALIGN_NEXT(transSize, 0x20),
                NULL);
     srcOffset += transSize;
     transLeft -= transSize;
