+++ src/JSystem/JKernel/JKRDvdRipper.cpp
@@ -50,7 +50,7 @@ void* JKRDvdRipper::loadToMainRAM(JKRDvdFile* dvdFile, u8* dst, JKRExpandSwitch 
     fileSizeAligned = ALIGN_NEXT(dvdFile->getFileSize(), 32);
     if (expandSwitch == EXPAND_SWITCH_UNKNOWN1) {
         u8 buffer[0x40];
-        u8 *bufPtr = (u8 *)ALIGN_NEXT((u32)buffer, 32);
+        u8 *bufPtr = (u8 *)ALIGN_NEXT((uintptr_t)buffer, 32);
         while (true) {
             int readBytes = DVDReadPrio(dvdFile->getFileInfo(), bufPtr, 0x20, 0, 2);
             if (readBytes >= 0) {
@@ -109,7 +109,7 @@ void* JKRDvdRipper::loadToMainRAM(JKRDvdFile* dvdFile, u8* dst, JKRExpandSwitch 
 
         if (offset != 0) {
             u8 buffer[0x40];
-            u8 *bufPtr = (u8 *)ALIGN_NEXT((u32)buffer, 32);
+            u8 *bufPtr = (u8 *)ALIGN_NEXT((uintptr_t)buffer, 32);
             while (true) {
                 int readBytes = DVDReadPrio(dvdFile->getFileInfo(), bufPtr, 32, (s32)offset, 2);
                 if (readBytes >= 0) {
@@ -419,7 +419,7 @@ static u8* nextSrcData(u8* src) {
     }
 
     memcpy(buf, src, limit);
-    u32 transSize = (u32)(szpEnd - (buf + limit));
+    u32 transSize = szpEnd - (buf + limit);
     if (transSize > transLeft) {
         transSize = transLeft;
     }
