+++ src/JSystem/JKernel/JKRHeap.cpp
@@ -95,7 +95,7 @@ bool JKRHeap::initArena(char** memory, u32* size, int maxHeaps) {
     OSSetArenaHi(ram_end);
 
     *memory = (char*)ram_start;
-    *size = (u32)ram_end - (u32)ram_start;
+    *size = (uintptr_t)ram_end - (uintptr_t)ram_start;
     return true;
 }
 
@@ -251,7 +251,7 @@ u8 JKRHeap::changeGroupID(u8 groupID) {
 
 /* 802B0918-802B0978       .text getMaxAllocatableSize__7JKRHeapFi */
 u32 JKRHeap::getMaxAllocatableSize(int alignment) {
-    u32 maxFreeBlock = (u32)getMaxFreeBlock();
+    uintptr_t maxFreeBlock = (uintptr_t)getMaxFreeBlock();
     u32 ptrOffset = (alignment - 1) & alignment - (maxFreeBlock & 0xf);
     return ~(alignment - 1) & (getFreeSize() - ptrOffset);
 }
@@ -291,7 +291,7 @@ JKRHeap* JKRHeap::find(void* memory) const {
 }
 
 /* 802B0A58-802B0AEC       .text dispose_subroutine__7JKRHeapFUlUl */
-void JKRHeap::dispose_subroutine(u32 begin, u32 end) {
+void JKRHeap::dispose_subroutine(uintptr_t begin, uintptr_t end) {
     JSUListIterator<JKRDisposer> last_iterator;
     JSUListIterator<JKRDisposer> next_iterator;
     JSUListIterator<JKRDisposer> iterator;
@@ -320,13 +320,13 @@ void JKRHeap::dispose_subroutine(u32 begin, u32 end) {
 
 /* 802B0AEC-802B0B14       .text dispose__7JKRHeapFPvUl */
 bool JKRHeap::dispose(void* ptr, u32 size) {
-    dispose_subroutine((u32)ptr, (u32)ptr + size);
+    dispose_subroutine((uintptr_t)ptr, (uintptr_t)ptr + size);
     return false;
 }
 
 /* 802B0B14-802B0B34       .text dispose__7JKRHeapFPvPv */
 void JKRHeap::dispose(void* begin, void* end) {
-    dispose_subroutine((u32)begin, (u32)end);
+    dispose_subroutine((uintptr_t)begin, (uintptr_t)end);
 }
 
 /* 802B0B34-802B0B8C       .text dispose__7JKRHeapFv */
