+++ src/JSystem/JKernel/JKRThread.cpp
@@ -44,7 +44,7 @@ JKRThread::JKRThread(u32 stack_size, int message_count, int param_3) : mThreadLi
 JKRThread::JKRThread(OSThread* thread, int message_count) : mThreadListLink(this) {
     mHeap = NULL;
     mThreadRecord = thread;
-    mStackSize = (u32)thread->stack_end - (u32)thread->stack_base;
+    mStackSize = (uintptr_t)thread->stack_end - (uintptr_t)thread->stack_base;
     mStackMemory = thread->stack_base;
 
     mMessageCount = message_count;
