+++ src/d/actor/d_a_bdk.cpp
@@ -3722,7 +3722,7 @@ static BOOL useHeapInit(fopAc_ac_c* i_actor) {
         }
     }
 
-    i_this->mpMorf->getModel()->setUserArea((u32)i_this);
+    i_this->mpMorf->getModel()->setUserArea((uintptr_t)i_this);
 
     i_this->mp2BC = mDoExt_J3DModel__create(i_this->mpMorf->getModel()->getModelData(), 0, 0x11020203);
     if (i_this->mp2BC == NULL) {
