+++ src/d/actor/d_a_ki.cpp
@@ -1274,7 +1274,7 @@ static BOOL useHeapInit(fopAc_ac_c* a_this) {
     }
 
     tex_anm_set(i_this, 1);
-    i_this->mpMorf->getModel()->setUserArea((u32)i_this);
+    i_this->mpMorf->getModel()->setUserArea((uintptr_t)i_this);
     return TRUE;
 }
