+++ src/d/actor/d_a_npc_mk.cpp
@@ -1715,7 +1715,7 @@ BOOL daNpc_Mk_c::CreateHeap() {
         }
     }
 
-    mpMorf->getModel()->setUserArea((u32)(this));
+    mpMorf->getModel()->setUserArea((uintptr_t)(this));
 
     mAcchCir.SetWall(60.0f, 30.0f);
     mObjAcch.Set(fopAcM_GetPosition_p(this), fopAcM_GetOldPosition_p(this), this, 1, &mAcchCir, fopAcM_GetSpeed_p(this));
