+++ src/d/actor/d_a_npc_photo.cpp
@@ -684,7 +684,7 @@ BOOL daNpcPhoto_c::createHeap() {
             modelData->getJointNodePointer(i)->setCallBack(daNpc_Photo_nodeCallBack);
         }
     }
-    mpMorf->getModel()->setUserArea((u32)this);
+    mpMorf->getModel()->setUserArea((uintptr_t)this);
     mAcchCir.SetWall(30.0f, 30.0f);
     mObjAcch.Set(fopAcM_GetPosition_p(this), fopAcM_GetOldPosition_p(this), this, 1, &mAcchCir, fopAcM_GetSpeed_p(this), fopAcM_GetAngle_p(this), fopAcM_GetShapeAngle_p(this));
     return TRUE;
