+++ src/d/actor/d_a_npc_tc.cpp
@@ -1891,7 +1891,7 @@ BOOL daNpc_Tc_c::_createHeap() {
     JUT_ASSERT(DEMO_SELECT(0xA38, 0xA42), m_jnt.getBackboneJntNum() >= 0);
     modelData->getJointNodePointer(TC_JNT_HEAD_e)->setCallBack(nodeCallBack);
     modelData->getJointNodePointer(TC_JNT_BACKBONE_e)->setCallBack(nodeCallBack);
-    mpMorf->getModel()->setUserArea(reinterpret_cast<u32>(this));
+    mpMorf->getModel()->setUserArea(reinterpret_cast<uintptr_t>(this));
 
     mTexPatternNum = 0;
     if (!initTexPatternAnm(false)) {
