+++ src/d/actor/d_a_npc_zl1.cpp
@@ -2978,7 +2978,7 @@ BOOL daNpc_Zl1_c::bodyCreateHeap() {
 
     mpMorf->getModel()->getModelData()->getJointNodePointer(m_hed_jnt_num & 0xffff)->setCallBack(nodeCB_Head);
     mpMorf->getModel()->getModelData()->getJointNodePointer(m_bbone_jnt_num & 0xffff)->setCallBack(nodeCB_BackBone);
-    mpMorf->getModel()->setUserArea((u32)(this));
+    mpMorf->getModel()->setUserArea((uintptr_t)(this));
     return TRUE;
 }
